refactor(BPU): separate history logic from bpu top module
